% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{apps_manual}
\alias{apps_manual}
\alias{apps_shinytest}
\alias{apps_shinyjster}
\alias{apps_testthat}
\alias{apps_deploy}
\title{Get names of Shiny apps to be tested}
\usage{
apps_manual(dir)

apps_shinytest(dir, suffix = "unknown")

apps_shinyjster(dir, browser = "unknown")

apps_testthat(dir)

apps_deploy(dir)
}
\arguments{
\item{dir}{base directory to look for shiny applications}

\item{suffix}{if a suffix string is provided, it will be appended to the shinytest flag used to search for apps that should not be tested with shinytest. If a null value or non character string is provided, all shinytest flags will be found.}

\item{browser}{if a browser string is provided, it will be appended to the shinyjster flag used to search for apps that should not be tested with shinyjster. If a null value or non character string is provided, all shinyjster flags will be found.}
}
\description{
All \code{apps_*} methods inspect each application to determine if if testing is possible.
}
\section{Functions}{
\itemize{
\item \code{apps_manual}: App folders that are to be manually tested. See \code{\link[=is_manual_app]{is_manual_app()}}.

\item \code{apps_shinytest}: App folders that contain a \verb{shinytest.R} file

\item \code{apps_shinyjster}: App folders that contain the text \code{shinyjster} in a Shiny R file

\item \code{apps_testthat}: App folders that contain a \verb{testthat.R} file

\item \code{apps_deploy}: App folders that contain a any Shiny app file
}}

