% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_status.R
\name{app_status}
\alias{app_status}
\alias{app_status_sheet}
\title{Testing status}
\usage{
app_status(dir = "apps", apps = apps_manual(dir))

app_status_sheet(
  dir = "apps",
  apps = apps_manual(dir),
  status = app_status(dir = dir, apps = apps)
)
}
\arguments{
\item{dir}{base folder containing all shiny apps}

\item{apps}{names of all shiny applications}

\item{status}{Status object to be printed.  If not supplied, it will be created used, \code{dir} and \code{apps}.}
}
\description{
Testing status
}
\section{Functions}{
\itemize{
\item \code{app_status}: Retrieve status of all manual executions

\item \code{app_status_sheet}: Display the google sheet status to be submitted
}}

\examples{
\dontrun{app_status()}
}
