% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{apps_shinytest}
\alias{apps_shinytest}
\alias{apps_shinyjster}
\alias{apps_testthat}
\alias{apps_manual}
\title{Get names of Shiny apps to be tested}
\usage{
apps_shinytest(dir)

apps_shinyjster(dir)

apps_testthat(dir)

apps_manual(dir)
}
\arguments{
\item{dir}{base directory to look for shiny applications}
}
\description{
All \code{apps_*} methods inspect each application to determine if if testing is possible.
}
