% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chart.R
\name{apps_shinytest}
\alias{apps_shinytest}
\alias{apps_shinyjster}
\title{Get names of apps to be tested with shinytest or shinyjster}
\usage{
apps_shinytest()

apps_shinyjster()
}
\description{
The data is fetched from a Google spreadsheet and cached for the duration
of the R session. To reset the cache, call \code{clear_cache()}.
}
