% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chart.R
\name{apps_testthat}
\alias{apps_testthat}
\title{Get names of apps to test with testthat/integration tests}
\usage{
apps_testthat(dir)
}
\arguments{
\item{dir}{Directory of apps to scan for testthat.R files}
}
\description{
Currently doesn't rely on the spreadsheet
}
