% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-ci.R
\name{assert_ci_output}
\alias{assert_ci_output}
\title{Assert \code{\link[=test_in_ci]{test_in_ci()}} output}
\usage{
assert_ci_output(ci_output)
}
\arguments{
\item{ci_output}{value received from \code{\link[=test_in_ci]{test_in_ci()}}}
}
\description{
Method called when \code{\link[=test_in_ci]{test_in_ci()}} is called with \code{assert = TRUE}.
}
