% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runtests.R
\name{assert_runtests}
\alias{assert_runtests}
\title{Assert \code{\link[=test_runtests]{test_runtests()}} output}
\usage{
assert_runtests(test_runtests_output)
}
\arguments{
\item{test_runtests_output}{value received from \code{\link[=test_runtests]{test_runtests()}}}
}
\description{
Method called when \code{\link[=test_runtests]{test_runtests()}} is called with \code{assert = TRUE}.
}
