% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-local.R
\name{assert_test_output}
\alias{assert_test_output}
\title{Assert \code{\link[=test_in_local]{test_in_local()}} output}
\usage{
assert_test_output(output)
}
\arguments{
\item{output}{value received from \code{\link[=test_in_local]{test_in_local()}}}
}
\description{
Method called when \code{\link[=test_in_local]{test_in_local()}} is called with \code{assert = TRUE}.
}
