% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_make_public.R
\name{connect_set_public}
\alias{connect_set_public}
\alias{connect_urls}
\title{Make Connect Shiny applications publically available}
\usage{
connect_set_public(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  account = "barret",
  server = "beta.rstudioconnect.com"
)

connect_urls(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  account = "barret",
  server = "beta.rstudioconnect.com"
)
}
\arguments{
\item{dir}{A base folder containing all applications to be deployed}

\item{apps}{A character vector of fully defined shiny application folders}

\item{account}{args supplied to \verb{[rsconnect::deployApp]}}

\item{server}{args supplied to \verb{[rsconnect::deployApp]}}
}
\description{
The corresponding \code{rsconnect} account should already exist before calling \code{connect_set_public}.  This can be done by calling \code{rsconnect::connectApiUser} to add the appropriate account information.
}
\section{Functions}{
\itemize{
\item \code{connect_set_public}: Set all the Shiny apps to be public on a Connect server using the Shiny applications provided in \verb{dir}

\item \code{connect_urls}: Retrieve the urls from a Connect server using the Shiny applications provided in \verb{dir}
}}

\examples{
\dontrun{
  rsconnect::addConnectServer(url = 'https://SERVER.com/API', name = 'CustomName')"
  rsconnect::connectApiUser('barret', 'CustomName', apiKey = 'SuperSecretKey')"
  deploy_apps(dir = 'apps', account = 'barret', server = 'CustomName')"
  connect_set_public(dir = 'apps', account = 'barret', server = 'CustomName')"
  urls <- connect_urls(dir = 'apps', account = 'barret', server = 'CustomName')
}
}
