% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_make_public.R
\name{connect_set_public}
\alias{connect_set_public}
\alias{connect_urls}
\title{Make Connect Shiny applications publically available}
\usage{
connect_set_public(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  account = "barret",
  server = "beta.rstudioconnect.com"
)

connect_urls(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  account = "barret",
  server = "beta.rstudioconnect.com"
)
}
\arguments{
\item{dir}{A base folder containing all applications to be deployed}

\item{apps}{A character vector of fully defined shiny application folders}

\item{account}{args supplied to \verb{[rsconnect::deployApp]}}

\item{server}{args supplied to \verb{[rsconnect::deployApp]}}

\item{api_key}{API key generated from the connect server.}
}
\description{
To not have to provide the key, the \code{rsconnect} account should already exist.  This can be done by calling \code{rsconnect::connectApiUser} to add the appropriate account information.
}
\examples{
\dontrun{
  connect_set_public(account = "barret", server = "beta.rstudioconnect.com")
}
}
