% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy_apps}
\alias{deploy_apps}
\title{Deploy apps to a server}
\usage{
deploy_apps(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  account = "testing-apps",
  server = "shinyapps.io",
  cores = 1,
  update_pkgs = c("all", "shinycoreci", "installed", "none"),
  retry = 3
)
}
\arguments{
\item{dir}{A base folder containing all applications to be deployed}

\item{apps}{A character vector of fully defined shiny application folders}

\item{account, server}{args supplied to \verb{[rsconnect::deployApp]}}

\item{cores}{number of cores to use when deploying}

\item{update_pkgs}{A logical value indicating if a check should be made to update all installed packages}

\item{retry}{If \code{TRUE}, try failure apps again. (Only happens once.)}
}
\description{
Run this in the terminal (not RStudio IDE) as it has issues when installing some packages.
}
