% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.R
\name{find_deps}
\alias{find_deps}
\title{Find package dependencies used in a directory}
\usage{
find_deps(dir = ".")
}
\arguments{
\item{dir}{The directory to look in.}
}
\description{
Find package dependencies used in a directory
}
