% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.R
\name{find_deps_installed}
\alias{find_deps_installed}
\title{Find package dependencies installed}
\usage{
find_deps_installed(libpath = shinyverse_libpath())
}
\description{
Find package dependencies installed
}
