% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{install_app_deps}
\alias{install_app_deps}
\alias{app_deps}
\title{Install app dependencies}
\usage{
install_app_deps(dir = "apps")

app_deps(dir = "apps")
}
\arguments{
\item{dir}{Directory containing apps}
}
\description{
Install app dependencies
}
