% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_exact.R
\name{install_exact_shinycoreci_deps}
\alias{install_exact_shinycoreci_deps}
\title{Install direct \pkg{shinycoreci} dependencies}
\usage{
install_exact_shinycoreci_deps(dir = "apps", try_again = TRUE)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{try_again}{Logical to determine if RStudio IDE should try again after a failure}
}
\description{
Install direct \pkg{shinycoreci} dependencies
}
