% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_suggested_pkgs}
\alias{load_suggested_pkgs}
\title{Load all suggested packages for session information}
\usage{
load_suggested_pkgs()
}
\description{
Load all suggested packages for session information
}
