% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-path.R
\name{resolve_libpath}
\alias{resolve_libpath}
\title{Resolve library path}
\usage{
resolve_libpath(..., local_pkgs = FALSE)
}
\arguments{
\item{local_pkgs}{If \code{TRUE}, local packages will be used instead of the isolated shinyverse installation.}
}
\description{
Resolve library path
}
\keyword{internal}
