% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspm.R
\name{rspm_install_scripts}
\alias{rspm_install_scripts}
\alias{rspm_pre_install_scripts}
\alias{rspm_all_install_scripts}
\title{RStudio Package Manager System Install Scripts}
\usage{
rspm_install_scripts(dir = "apps", distro = "bionic")

rspm_pre_install_scripts(dir = "apps", distro = "bionic")

rspm_all_install_scripts(dir = "apps", distro = "bionic")
}
\arguments{
\item{dir}{base directory to look for shiny applications}

\item{distro}{Docker distro to use. Such as \verb{'bionic'} or \verb{'centos7'}}
}
\description{
RStudio Package Manager System Install Scripts
}
