% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_runtests.R
\name{test_runtests}
\alias{test_runtests}
\alias{test_shinytest}
\alias{test_shinyjster}
\alias{test_testthat}
\title{Test apps using \code{shiny::runTests()}}
\usage{
test_runtests(
  dir = "apps",
  apps = apps_runtests(dir, filter = filter),
  filter = NULL,
  assert = TRUE,
  timeout = as.difftime(10, units = "mins"),
  retries = 2,
  update_pkgs = TRUE,
  update_app_pkgs = TRUE
)

test_shinytest(
  dir = "apps",
  apps = apps_shinytest(dir),
  assert = TRUE,
  retries = 3
)

test_shinyjster(
  dir = "apps",
  apps = apps_shinyjster(dir),
  assert = TRUE,
  retries = 3
)

test_testthat(
  dir = "apps",
  apps = apps_testthat(dir),
  assert = TRUE,
  retries = 3
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{filter}{filter to use on test file within application}

\item{assert}{logical value which will determine if \code{\link[=assert_runtests]{assert_runtests()}} will be called on the result}

\item{timeout}{Length of time allowed for an application's full test suit can run before determining it is a failure}

\item{retries}{number of attempts to retry before declaring the test a failure}

\item{update_pkgs}{Logical value which will try to install all required shiny packages used for testing}

\item{update_app_pkgs}{Logical value which will try to install all required app packages used for testing}
}
\description{
Test apps using \code{shiny::runTests()}
}
\section{Functions}{
\itemize{
\item \code{test_runtests}: Generic method to call all testing files

\item \code{test_shinytest}: Only execute shinytest.R test files

\item \code{test_shinyjster}: Only execute shinyjster.R test files

\item \code{test_testthat}: Only execute testtat.R test files
}}

