% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-path.R
\name{shinycoreci_libpath}
\alias{shinycoreci_libpath}
\alias{shinycoreci_clean_libpaths}
\title{Shinyverse libpath}
\usage{
shinycoreci_libpath()

shinycoreci_clean_libpaths()
}
\description{
Methods to get and reset the shinyverse \code{libpath}.
}
\section{Functions}{
\itemize{
\item \code{shinycoreci_libpath()}: Library path that will persist across installations. But will have a different path for different R versions

\item \code{shinycoreci_clean_libpaths()}: Removes the cached R library

}}
