% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-path.R
\name{shinyverse_libpath}
\alias{shinyverse_libpath}
\alias{shinyverse_clean_libpath}
\title{Shinyverse libpath}
\usage{
shinyverse_libpath()

shinyverse_clean_libpath()
}
\description{
Methods to get and reset the shinyverse \code{libpath}.
}
\section{Functions}{
\itemize{
\item \code{shinyverse_libpath}: Library path that will persist across installations. But will have a different path for different R versions

\item \code{shinyverse_clean_libpath}: Removes the cached R library
}}

