% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{save_test_results}
\alias{save_test_results}
\alias{view_test_results}
\title{View (and save) \code{test_runtests()} output}
\usage{
save_test_results(test_runtests_output, gha_branch_name, pr_number, username)

view_test_results(
  dir = "apps",
  update = TRUE,
  from = Sys.Date() - 5,
  to = Sys.Date()
)
}
\arguments{
\item{test_runtests_output}{output from \code{test_runtests()}.}

\item{gha_branch_name}{an identifier determined by the GHA workflow}

\item{pr_number}{the pull request number.}

\item{username}{the username of the github profile.}

\item{dir}{base folder to look for applications}

\item{update}{whether or not to fetch the latest test results}

\item{from}{start date.}

\item{to}{end date.}
}
\description{
Use \code{view_test_results()} to preview test results generated by \code{save_test_results()}
(saving should only ever be done in a GHA workflow).
}
