% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-browser.R
\name{test_in_browser}
\alias{test_in_browser}
\title{Test apps within the terminal}
\usage{
test_in_browser(
  app_name = apps[1],
  apps = apps_manual,
  port = 8080,
  port_background = NULL,
  host = "127.0.0.1",
  local_pkgs = FALSE
)
}
\arguments{
\item{port_background}{\code{port} for the background app process}

\item{local_pkgs}{If \code{TRUE}, local packages will be used instead of the isolated shinyverse installation.}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}
}
\description{
Automatically runs the next app in a fresh callr::r_bg session.  To stop, close the shiny application window.
}
\examples{
\dontrun{test_in_browser()}
}
