% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-ide.R
\name{test_in_ide}
\alias{test_in_ide}
\title{Test apps within RStudio IDE}
\usage{
test_in_ide(
  app_name = apps[1],
  apps = apps_manual,
  ...,
  port = 8000,
  host = "127.0.0.1",
  delay = 1,
  local_pkgs = FALSE,
  viewer = NULL,
  refresh_ = FALSE
)
}
\arguments{
\item{delay}{Time to wait between applications. [\code{1}]}

\item{local_pkgs}{If \code{TRUE}, local packages will be used instead of the isolated shinyverse installation.}

\item{viewer}{RStudio IDE viewer to use.  [\code{"pane"}]}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}
}
\description{
Automatically runs the next app in a fresh RStudio session after closing the current app. To stop,  send an interrupt signal (\verb{esc} or \verb{ctrl+c}) to the app twice in rapid succession.
}
\details{
Kill testing by hitting \verb{esc} in RStudio.

If \code{options()} need to be set, set them in your \preformatted{.Rprofile} file.  See \code{usethis::edit_r_profile()}
}
\examples{
\dontrun{test_in_ide(dir = "apps")}
}
