% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-local.R
\name{test_in_local}
\alias{test_in_local}
\title{Test apps using \code{shiny::runTests()} using local libpath}
\usage{
test_in_local(
  apps = apps_with_tests(repo_dir),
  ...,
  assert = TRUE,
  timeout = 10 * 60,
  retries = 2,
  repo_dir = rprojroot::find_package_root_file(),
  local_pkgs = FALSE
)
}
\arguments{
\item{apps}{applications within \verb{dir} to run}

\item{...}{ignored}

\item{assert}{logical value which will determine if \code{\link[=assert_test_output]{assert_test_output()}} will be called on the result}

\item{timeout}{Length of time allowed for an application's full test suit can run before determining it is a failure}

\item{retries}{number of attempts to retry before declaring the test a failure}

\item{repo_dir}{Location of local shinycoreci repo}

\item{local_pkgs}{If \code{TRUE}, the local \code{.libPaths()[1]} will be used. If \code{FALSE}, a persistent shinycoreci libpath will be used. Only when \code{local_pkgs=FALSE} will app dependencies be installed.}
}
\description{
Test apps using \code{shiny::runTests()} using local libpath
}
