% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shinyjster.R
\name{test_shinyjster}
\alias{test_shinyjster}
\title{Test shinyjster}
\usage{
test_shinyjster(
  dir = "apps",
  apps = apps_shinyjster(dir),
  port = 8000,
  host = "127.0.0.1",
  debug_port = NULL,
  browser = c("chrome", "firefox"),
  type = c("serial", "lapply", "parallel", "callr"),
  assert = TRUE
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}
}
\description{
Test shinyjster
}
