% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_shinyjster_app}
\alias{test_shinyjster_app}
\title{Test shinyjster app}
\usage{
test_shinyjster_app(
  name,
  browser_func = switch(name, chrome = shinyjster::selenium_chrome(timeout = timeout,
    dimensions = dimensions), firefox = shinyjster::selenium_firefox(timeout = timeout,
    dimensions = dimensions), edge = shinyjster::selenium_edge(timeout = timeout,
    dimensions = dimensions), ie = shinyjster::selenium_ie(timeout = timeout, dimensions
    = dimensions), {     message("unknown browser name supplied. Using 'chrome'")    
    shinyjster::selenium_chrome() }),
  apps = "..",
  timeout = 2 * 60,
  dimensions = "1200x1200"
)
}
\arguments{
\item{name}{browser name to use to determine the browser_func}

\item{browser_func}{One of \code{c("chrome", "firefox", "edge", "ie")}. If an unknown \code{browser_name} is used, \code{browser_func} will default to using \code{chrome}.}

\item{apps}{shiny app to be tested using shinyjster}

\item{timeout}{Number of seconds before selenium closes the browser}

\item{dimensions}{A string in the form of \verb{"WIDTHxHEIGHT"}. Ex: \code{"1800x1200"}}
}
\description{
Method to avoid editing MANY different files to do the same default behavior of \code{shinyjster::test_jster(browsers = selenium_chrome(), type = "lapply")}.
}
\details{
Some extra logic wraps around whether or not a browser is available on the given platform.
}
\seealso{
\code{\link[=test_shinytest_app]{test_shinytest_app()}} and \code{\link[=test_runtests]{test_runtests()}}
}
