% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_shinytest_app}
\alias{test_shinytest_app}
\title{Test shinytest app}
\usage{
test_shinytest_app(appDir = "..", suffix = platform_rversion(), ...)
}
\description{
Test shinytest app
}
\seealso{
\code{\link[=test_shinyjster_app]{test_shinyjster_app()}}, \code{\link[=test_runtests]{test_runtests()}}, and \code{\link[=platform_rversion]{platform_rversion()}}
}
