% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_testhat_app}
\alias{test_testhat_app}
\title{Test shinytest app}
\usage{
test_testhat_app()
}
\description{
Test shinytest app
}
\seealso{
\code{\link[=test_shinyjster_app]{test_shinyjster_app()}} and \code{\link[=test_runtests]{test_runtests()}}
}
