% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_testthat.R
\name{test_testthat}
\alias{test_testthat}
\title{Test testthat}
\usage{
test_testthat(dir = "apps", apps = apps_testthat(dir), suffix = platform())
}
\arguments{
\item{dir}{Name of directory to look for shiny apps, and to save platform
information.}

\item{apps}{Character vector of shiny applications}

\item{suffix}{Suffix to use for sysinfo file}
}
\description{
Test testthat
}
