% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trigger}
\alias{trigger}
\title{Generate a repository event.}
\usage{
trigger(
  event_type,
  repo = "rstudio/shinycoreci-apps",
  client_payload = list(),
  auth_token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{event_type}{The name of the event to create on the repository}

\item{repo}{The GitHub repo to create the event on; defaults to
rstudio/shinycoreci-apps}

\item{client_payload}{The JSON object to make available in the workflow as
the \code{github.event.client_payload} object}

\item{auth_token}{Your GitHub OAuth2 token; defaults to
\code{Sys.getenv("GITHUB_PAT")}}
}
\description{
This function uses the GitHub API to create a \href{https://developer.github.com/v3/repos/#create-a-repository-dispatch-event}{repository dispatch event}
that can trigger workflows. Currently, the \code{testthat.yml} workflow
registers itself for the \code{shinytest-apps} event, and so can be initiated by
running this function with an \code{event_type} of "shinytest-apps".
}
