% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_methods.R
\name{use_manual_app}
\alias{use_manual_app}
\title{Flag an app to be manually tested}
\usage{
use_manual_app(app_dir)
}
\arguments{
\item{app_dir}{Shiny application directory containing an app.R, ui.R, server.R, or index.Rmd}
}
\description{
All \code{apps_*} methods inspect each application to determine if if testing is possible.
}
