% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_methods.R
\name{use_shinyjster}
\alias{use_shinyjster}
\title{Create Shinyjster test file}
\usage{
use_shinyjster(app_dir)
}
\arguments{
\item{app_dir}{Location of shiny application to test}
}
\description{
This creates a testing file that will test shinyjster on each applicable browser.
}
