% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_methods.R
\name{use_tests_shinyjster}
\alias{use_tests_shinyjster}
\title{Create Shinyjster test file}
\usage{
use_tests_shinyjster(app_dir)
}
\arguments{
\item{app_dir}{Location of shiny application to test}
}
\description{
This creates a testing file to be used with \code{\link[=test_runtests]{test_runtests()}}.  It will create a file for each browser.
}
\seealso{
\code{\link[=test_jster]{test_jster()}}
}
