% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_methods.R
\name{use_tests_shinytest}
\alias{use_tests_shinytest}
\title{Create Shinyjster test file}
\usage{
use_tests_shinytest(app_dir)
}
\arguments{
\item{app_dir}{Location of shiny application to test}
}
\description{
Create Shinyjster test file
}
\seealso{
\code{\link[=test_shinytest]{test_shinytest()}}
}
