% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_methods.R
\name{use_tests_testthat}
\alias{use_tests_testthat}
\title{Create an app testthat test file}
\usage{
use_tests_testthat(app_dir)
}
\arguments{
\item{app_dir}{Location of shiny application to test}
}
\description{
Create an app testthat test file
}
\seealso{
\code{\link[=test_testthat]{test_testthat()}}
}
