% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_all_test_diff.R
\name{view_all_test_diff}
\alias{view_all_test_diff}
\title{View Shinytest Diff}
\usage{
view_all_test_diff(
  dir = "apps",
  sha = git_sha(dir),
  ask = interactive(),
  ...,
  repo_dir = file.path(dir, "..")
)
}
\arguments{
\item{dir}{Root app folder path}

\item{sha}{git sha of base branch to look for}

\item{ask}{Logical which allows for particular branches to be inspected}

\item{...}{Extra arguments passed to \code{shinytest::viewTestDiff}}

\item{repo_dir}{Root repo folder path}
}
\description{
View Shinytest Diff
}
