% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_test_diff.R
\name{view_test_diff}
\alias{view_test_diff}
\title{View Shinytest Diff}
\usage{
view_test_diff(suffix = platform_rversion(), path = "apps", ...)
}
\arguments{
\item{suffix}{Test output suffix to compare against}

\item{path}{Root folder path}

\item{...}{Extra arguments passed to \code{shinytest::viewTestDiff}}
}
\description{
View Shinytest Diff
}
