% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_test_images.R
\name{view_test_images}
\alias{view_test_images}
\title{View Shinytest Images}
\usage{
view_test_images(dir = "apps")
}
\arguments{
\item{dir}{path pointing to shinycoreci-apps/apps}
}
\description{
View Shinytest Images
}
