% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_options}
\alias{with_options}
\title{Eval with Options}
\usage{
with_options(new_options, code)
}
\arguments{
\item{new_options}{list of options to be supplied to \code{options()}}

\item{code}{code to evaluate, given the new options}
}
\description{
Eval with Options
}
\examples{
\dontrun{with_options(list(warn = 0), {
  warning('just a warning')
  with_options(list(warn = 2), { warning('made into error') })
  warning('just a warning')
})}
}
