% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.R
\name{write_sysinfo}
\alias{write_sysinfo}
\title{Write system information to a file}
\usage{
write_sysinfo(file = stdout(), libpath = shinyverse_libpath())
}
\arguments{
\item{file}{Name of file, or file object to write to (defaults to stdout).}

\item{libpath}{Library path to find installaed packages.}
}
\description{
Write system information to a file
}
