% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accept_all_snaps.R
\name{accept_snaps}
\alias{accept_snaps}
\title{Accept all snapshot changes}
\usage{
accept_snaps(repo_dir = ".")
}
\arguments{
\item{repo_dir}{Root repo folder path}
}
\description{
Accepts \strong{all} snapshot changes for every shiny application in \code{inst/apps}.
}
