% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-connect-urls.R
\name{connect_set_public}
\alias{connect_set_public}
\alias{connect_urls}
\title{Make Connect Shiny applications publically available}
\usage{
connect_set_public(
  apps = apps_shiny,
  account = "barret",
  server = "beta.rstudioconnect.com"
)

connect_urls(
  apps = apps_deploy,
  account = "barret",
  server = "beta.rstudioconnect.com"
)
}
\arguments{
\item{apps}{A character vector of fully defined shiny application folders}

\item{account, server}{args supplied to \verb{[rsconnect::deployApp]}}
}
\description{
The corresponding \code{rsconnect} account should already exist before calling \code{connect_set_public}.  This can be done by calling \code{rsconnect::connectApiUser} to add the appropriate account information.
}
\section{Functions}{
\itemize{
\item \code{connect_set_public}: Set all the Shiny apps to be public on a Connect server

\item \code{connect_urls}: Retrieve the urls from a Connect server using the Shiny applications provided in \verb{dir}
}}

\examples{
\dontrun{
  rsconnect::addConnectServer(url = 'https://SERVER.com/API', name = 'CustomName')
  rsconnect::connectApiUser('barret', 'CustomName', apiKey = 'SuperSecretKey')
  deploy_apps(account = 'barret', server = 'CustomName')
  connect_set_public(account = 'barret', server = 'CustomName')
  urls <- connect_urls(account = 'barret', server = 'CustomName')
}
}
