% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy-apps.R
\name{deploy_apps}
\alias{deploy_apps}
\title{Deploy apps to a server}
\usage{
deploy_apps(
  apps = apps_shiny,
  account = "testing-apps",
  server = "shinyapps.io",
  ...,
  extra_packages = NULL,
  cores = 1,
  retry = 2,
  retrying_ = FALSE
)
}
\arguments{
\item{apps}{A character vector of fully defined shiny application folders}

\item{account, server}{args supplied to \verb{[rsconnect::deployApp]}}

\item{cores}{number of cores to use when deploying}

\item{retry}{If \code{TRUE}, try failure apps again. (Only happens once.)}

\item{dir}{A base folder containing all applications to be deployed}

\item{update_pkgs}{A logical value indicating if a check should be made to update all installed packages}
}
\description{
Run this in the terminal (not RStudio IDE) as it has issues when installing some packages.
}
