% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{docker_run_sso}
\alias{docker_run_sso}
\alias{docker_run_ssp}
\title{Docker Testing}
\usage{
docker_run_sso(
  release = c("focal", "bionic", "centos7"),
  port = switch(release, centos7 = 7878, 3838),
  r_version = c("4.1", "4.0", "3.6", "3.5"),
  tag = NULL,
  launch_browser = TRUE
)

docker_run_ssp(
  release = c("focal", "bionic", "centos7"),
  port = switch(release, centos7 = 8989, 4949),
  r_version = c("4.1", "4.0", "3.6", "3.5"),
  tag = NULL,
  launch_browser = TRUE
)
}
\arguments{
\item{release}{Distro release name, such as "focal" for ubuntu or "7" for centos}

\item{port}{port to have server function locally}

\item{r_version}{R version to use. Ex: \code{"3.6"}}

\item{tag}{Extra tag information for the docker image. This will prepend a \verb{-} if a value is given.}

\item{launch_browser}{Logical variable that determines if the browser should open to the specified port location}
}
\description{
Docker Testing
}
\section{Functions}{
\itemize{
\item \code{docker_run_sso}: Run SSO in a docker container

\item \code{docker_run_ssp}: Run SSP in a docker container
}}

