% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_all_gha_branches.R
\name{fix_all_gha_branches}
\alias{fix_all_gha_branches}
\title{Fix All Failing Shinytest Diffs and Merge All \verb{gha-} Branches}
\usage{
fix_all_gha_branches(
  dir = "apps",
  sha = git_sha(dir),
  ...,
  save_results = NULL,
  ask_apps = FALSE,
  ask_branches = TRUE,
  ask_if_not_main = TRUE,
  accept_all = FALSE,
  repo_dir = file.path(dir, "..")
)
}
\arguments{
\item{dir}{Root app folder path}

\item{sha}{git sha of base branch to look for}

\item{...}{Extra arguments passed to \code{shinytest::viewTestDiff}}

\item{save_results}{Logical which commits and merges all branches after viewing test results}

\item{ask_apps, ask_branches}{Logical which allows for particular apps branches to be inspected}

\item{ask_if_not_main}{Logical which will check if \code{main} is the base branch}

\item{accept_all}{Logical which will accept all changes for the apps selected when \code{ask_apps = TRUE}. This parameter is only recommended for use when the shinytest differences are known and can not be displayed using the shiny application.}

\item{repo_dir}{Root repo folder path}
}
\description{
Function to walk the user through merging all \verb{gha-} branches.  This function will NOT push to the repository. Users must \verb{git push} themselves.
}
\details{
This function will NOT fix \code{shinyjster} failures.

Outline of steps performed:
\enumerate{
\item Validate there are no git changes or untracked files in the current base branch
\item Validate there are no unexpected shinytest folders
\item If \code{isTRUE(ask)}, ask for which branches to work with
\item Find all apps to update
\item For each failing shinytest app, accept / reject shinytest changes in each \verb{gha-} branch. Add an individualized commit message including the app and environment flavor information
\item For each \verb{gha-} branch, validate there are no unexpected shinytest folders. (Should be no unexpected folders after accepting / rejecting apps.)
\item For each \verb{gha-} branch, merge into the base branch.
\item For each \verb{gha-} branch, delete the locally checked out \verb{gha-} branch. (Cleans up the local repo.)
\item Tell the user to call \verb{git push}.
}
}
