% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{install_app_deps}
\alias{install_app_deps}
\alias{app_deps}
\title{Install app dependencies}
\usage{
install_app_deps(
  dir = "apps",
  apps = apps_runtests(dir),
  include_shinycoreci = TRUE
)

app_deps(dir = "apps", apps = apps_runtests(dir), include_shinycoreci = TRUE)
}
\arguments{
\item{dir}{Directory containing apps}

\item{apps}{apps to find the dependencies of}

\item{include_shinycoreci}{Logical value which determines if shinycoreci should be included}
}
\description{
Install app dependencies
}
