% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps_exact.R
\name{install_exact_shinycoreci_deps}
\alias{install_exact_shinycoreci_deps}
\title{Install direct \pkg{shinycoreci} dependencies}
\usage{
install_exact_shinycoreci_deps(
  dir = "apps",
  apps = apps_runtests(dir),
  try_again = TRUE,
  assert = interactive()
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{try_again}{Logical to determine if RStudio IDE should try again after a failure}

\item{assert}{Logical to determine if an error should be thrown or RStudio should be restarted}
}
\description{
If something goes wrong, default to installing \code{shinycoreci} from GitHub. Then try installing the dependencies again. See the examples.
}
\examples{
\dontrun{remotes::install_github("rstudio/shinycoreci")
shinycoreci::install_exact_shinycoreci_deps("apps")}
}
