% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{platform_rversion}
\alias{platform_rversion}
\title{Platform and R Version}
\usage{
platform_rversion(platform_val = platform(), r_version = r_version_short())
}
\arguments{
\item{r_version}{See \code{\link[=r_version_short]{r_version_short()}}}

\item{platform}{See \code{\link[=platform]{platform()}}}
}
\description{
Platform and R Version
}
