% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rspm.R
\name{rspm_install_scripts}
\alias{rspm_install_scripts}
\alias{rspm_pre_install_scripts}
\alias{rspm_all_install_scripts}
\title{RStudio Package Manager System Install Scripts}
\usage{
rspm_install_scripts(
  dir = "apps",
  release = c("focal", "bionic", "xenial", "centos7")
)

rspm_pre_install_scripts(
  dir = "apps",
  release = c("focal", "bionic", "xenial", "centos7")
)

rspm_all_install_scripts(
  dir = "apps",
  release = c("focal", "bionic", "xenial", "centos7")
)
}
\arguments{
\item{dir}{base directory to look for shiny applications}

\item{release}{Docker release to use. Such as \verb{'focal'} or \verb{'centos7'}}
}
\description{
RStudio Package Manager System Install Scripts
}
\section{Functions}{
\itemize{
\item \code{rspm_install_scripts}: Install script needed for R packages to run

\item \code{rspm_pre_install_scripts}: Install script needed for R packages to install

\item \code{rspm_all_install_scripts}: All install script needed for R packages
}}

