% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-ci.R
\name{test_in_ci}
\alias{test_in_ci}
\title{Test apps using \code{shiny::runTests()}}
\usage{
test_in_ci(
  apps = apps_tests,
  assert = TRUE,
  timeout = 10 * 60,
  retries = 2,
  repo_dir = "."
)
}
\arguments{
\item{apps}{applications within \verb{dir} to run}

\item{assert}{logical value which will determine if \code{\link[=assert_ci_output]{assert_ci_output()}} will be called on the result}

\item{timeout}{Length of time allowed for an application's full test suit can run before determining it is a failure}

\item{retries}{number of attempts to retry before declaring the test a failure}

\item{repo_dir}{Location of local shinycoreci repo}
}
\description{
Test apps using \code{shiny::runTests()}
}
\section{Functions}{
\itemize{
\item \code{test_in_ci}: Generic method to call all testing files
}}

