% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-test-results.R
\name{save_test_results}
\alias{save_test_results}
\alias{view_test_results}
\title{View (and save) \code{test_in_ci()} output}
\usage{
save_test_results(
  ci_output,
  gha_branch_name,
  pr_number,
  username,
  repo_dir = "."
)

view_test_results(repo_dir = ".")
}
\arguments{
\item{ci_output}{output from \code{test_in_ci()}.}

\item{gha_branch_name}{an identifier determined by the GHA workflow}

\item{pr_number}{the pull request number.}

\item{username}{the username of the github profile.}

\item{repo_dir}{Location of local shinycoreci repo}
}
\description{
Use \code{view_test_results()} to preview test results generated by \code{save_test_results()}
(saving should only ever be done in a GHA workflow).
}
