% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_results.R
\name{save_test_results}
\alias{save_test_results}
\title{View (and save) \code{test_runtests()} output}
\usage{
save_test_results(test_runtests_output, gha_branch_name, pr_number, username)
}
\arguments{
\item{test_runtests_output}{output from \code{test_runtests()}.}

\item{gha_branch_name}{an identifier determined by the GHA workflow}

\item{pr_number}{the pull request number.}

\item{username}{the username of the github profile.}
}
\description{
Use \code{view_test_results()} to preview test results generated by \code{save_test_results()}
(saving should only ever be done in a GHA workflow).
}
