% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_in_browser.R
\name{test_in_browser}
\alias{test_in_browser}
\title{Test apps within the terminal}
\usage{
test_in_browser(
  dir = "apps",
  apps = basename(apps_manual(dir)),
  app = apps[1],
  port = 8080,
  port_background = NULL,
  host = "127.0.0.1",
  update_pkgs = TRUE,
  verify = TRUE
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{port_background}{\code{port} for the background app process}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{update_pkgs}{Logical that will try to automatically install packages. [\code{TRUE}]}

\item{verify}{Logical that will try to confirm shinycoreci-apps directory is the master branch}
}
\description{
Automatically runs the next app in a fresh callr::r_bg session.  To stop, close the shiny application window.
}
\examples{
\dontrun{test_in_browser(dir = "apps")}
}
