% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_in_connect.R, R/test_in_shinyapps_io.R
\name{test_in_connect}
\alias{test_in_connect}
\alias{test_in_shinyappsio}
\title{Test deployed apps}
\usage{
test_in_connect(
  dir = "apps",
  urls = test_in_connect_urls(),
  apps = names(urls),
  app = apps[1],
  server = attr(urls, "server"),
  account = attr(urls, "account"),
  port = 8080,
  host = "127.0.0.1"
)

test_in_shinyappsio(
  dir = "apps",
  apps = basename(apps_deploy(dir)),
  app = apps[1],
  port = 8080,
  host = "127.0.0.1",
  account = "testing-apps",
  server = "shinyapps.io"
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{urls}{Named vector of urls to visit. This should be the output of \verb{[connect_urls]}. By default this is determined using the \code{server}, \code{account}, \code{dir}, and \code{apps}}

\item{apps}{applications within \verb{dir} to run}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{server, account}{Parameters that could be supplied to \verb{[rsconnect::deployApp]}}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
}
\description{
Automatically runs the next app in a fresh callr::r_bg session.  To stop, close the shiny application window.
}
\section{Functions}{
\itemize{
\item \code{test_in_connect}: Test connect applications given the server and account

\item \code{test_in_shinyappsio}: Test connect applications given the server and account
}}

\examples{
\dontrun{test_in_connect(dir = "apps")}
\dontrun{test_in_test_in_shinyapps_io(dir = "apps")}
}
