% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-in-connect.R, R/test-in-shinyapps_io.R
\name{test_in_connect}
\alias{test_in_connect}
\alias{test_in_shinyappsio}
\title{Test deployed apps}
\usage{
test_in_connect(
  app_name = apps[1],
  apps = apps_manual,
  ...,
  urls = default_connect_urls,
  server = attr(urls, "server"),
  account = attr(urls, "account"),
  port = 8080,
  host = "127.0.0.1"
)

test_in_shinyappsio(
  app_name = apps[1],
  apps = apps_manual,
  port = 8080,
  host = "127.0.0.1",
  account = "testing-apps",
  server = "shinyapps.io"
)
}
\arguments{
\item{app_name}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{apps}{List of apps to test}

\item{...}{ignored}

\item{urls}{Named vector of urls to visit. This should be the output of \verb{[connect_urls]}. By default this will set \code{server}, \code{account}, \code{dir}, and \code{apps}}

\item{server, account}{Parameters that could be supplied to \verb{[rsconnect::deployApp]}}

\item{port}{\code{port} for the foreground app process}

\item{host}{\code{host} for the foreground app processes}
}
\description{
Automatically runs the next app in a fresh callr::r_bg session.  To stop, close the shiny application window.
}
\section{Functions}{
\itemize{
\item \code{test_in_connect()}: Test connect applications given the server and account

\item \code{test_in_shinyappsio()}: Test connect applications given the server and account

}}
\examples{
\dontrun{test_in_connect(dir = "apps")}
\dontrun{test_in_test_in_shinyapps_io(dir = "apps")}
}
