% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_in_ide.R
\name{test_in_ide}
\alias{test_in_ide}
\title{Test apps within RStudio IDE}
\usage{
test_in_ide(
  dir = "apps",
  apps = basename(apps_manual(dir)),
  app = apps[1],
  port = 8000,
  host = "127.0.0.1",
  delay = 1,
  update_pkgs = TRUE,
  viewer = NULL,
  verify = TRUE
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{delay}{Time to wait between applications. [\code{1}]}

\item{update_pkgs}{Logical that will try to automatically install packages. [\code{TRUE}]}

\item{viewer}{RStudio IDE viewer to use.  [\code{"pane"}]}

\item{verify}{Logical that will try to confirm shinycoreci-apps directory is the master branch}
}
\description{
Automatically runs the next app in a fresh RStudio session after closing the current app. To stop,  send an interrupt signal (\verb{esc} or \verb{ctrl+c}) to the app twice in rapid succession.
}
\details{
Kill testing by hitting \verb{esc} in RStudio.

If \code{options()} need to be set, set them in your \preformatted{.Rprofile} file.  See \code{usethis::edit_r_profile()}
}
\examples{
\dontrun{test_in_ide(dir = "apps")}
}
