% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_in_ide.R
\name{test_in_ide}
\alias{test_in_ide}
\title{Test apps within RStudio IDE}
\usage{
test_in_ide(
  dir = "apps",
  apps = apps_manual(dir),
  app = apps[1],
  port = 8000,
  host = "127.0.0.1",
  delay = 1,
  update_pkgs = TRUE,
  viewer = NULL,
  verify = TRUE
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port between 3000:8000, excluding ports that are blocked
by Google Chrome for being considered unsafe: 3659, 4045, 5060,
5061, 6000, 6566, 6665:6669 and 6697. Up to twenty random
ports will be tried.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{delay}{Time to wait between applications. [\code{1}]}

\item{update_pkgs}{Logical value which will try to install all required shiny packages used for testing}

\item{viewer}{RStudio IDE viewer to use.  [\code{"pane"}]}

\item{verify}{Logical that will try to confirm shinycoreci-apps directory is the main branch}
}
\description{
Automatically runs the next app in a fresh RStudio session after closing the current app. To stop,  send an interrupt signal (\verb{esc} or \verb{ctrl+c}) to the app twice in rapid succession.
}
\details{
Kill testing by hitting \verb{esc} in RStudio.

If \code{options()} need to be set, set them in your \preformatted{.Rprofile} file.  See \code{usethis::edit_r_profile()}
}
\examples{
\dontrun{test_in_ide(dir = "apps")}
}
