% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_in_sso.R
\name{test_in_sso}
\alias{test_in_sso}
\alias{test_in_ssp}
\title{Test Apps in SSO/SSP}
\usage{
test_in_sso(
  dir = "apps",
  apps = basename(apps_sso(dir)),
  app = apps[1],
  release = c("bionic", "xenial", "centos7"),
  r_version = c("3.6", "3.5"),
  tag = NULL,
  port = 8080,
  port_background = switch(release, centos7 = 7878, 3838),
  host = "127.0.0.1"
)

test_in_ssp(
  dir = "apps",
  apps = basename(apps_ssp(dir)),
  app = apps[1],
  release = c("bionic", "xenial", "centos7"),
  r_version = c("3.6", "3.5"),
  tag = NULL,
  port = 8080,
  port_background = switch(release, centos7 = 8989, 4949),
  host = "127.0.0.1"
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{app}{app number or name to start with. If numeric, it will match the leading number in the testing application}

\item{release}{Distro release name, such as "bionic" for ubuntu or "7" for centos}

\item{r_version}{R version to use. Ex: \code{"3.6"}}

\item{tag}{Extra tag information for the docker image. This will prepend a \verb{-} if a value is given.}

\item{port}{Port for shiny application}

\item{port_background}{Port to connect to the Docker container}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}
}
\description{
Automatically launches docker in a background process.  Once the docker is ready, a shiny application will be launched to help move through the applications.
}
\details{
The docker application will stop when the shiny application exits.
}
\section{Functions}{
\itemize{
\item \code{test_in_sso}: Test SSO Shiny applications

\item \code{test_in_ssp}: Test SSP Shiny applications
}}

\examples{
\dontrun{test_in_connect(dir = "apps")}
}
