% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shinyjster.R
\name{test_shinyjster}
\alias{test_shinyjster}
\title{Test shinyjster}
\usage{
test_shinyjster(
  dir = "apps",
  apps = apps_shinyjster(),
  port = 8000,
  host = "127.0.0.1",
  debug_port = NULL,
  browser = c("chrome", "firefox"),
  type = c("serial", "lapply", "parallel", "callr"),
  assert = TRUE
)
}
\arguments{
\item{apps}{applications within \verb{dir} to run}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{debug_port}{Port for chrome debugger}

\item{browser}{Browser to be set for testing.  If using \code{"chrome"} or \code{"firefox"}, the system application will be used.}

\item{type}{Single value to determine how applications are executed. \describe{
\item{\code{'parallel'}}{Runs apps using \code{parallel::mclapply} using \code{cores} cores}
\item{\code{'callr'}}{Runs apps using \code{callr::r_bg} using \code{cores} cores}
\item{\code{'serial'}}{Runs apps one after another using \code{lapply}. \code{port} is only used with options \code{'serial'} and \code{'lapply'}}
}}

\item{assert}{Logical which determines if \link[shinyjster]{assert_jster} should be run on the final output}
}
\description{
Test shinyjster
}
