% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shinyjster.R
\name{test_shinyjster}
\alias{test_shinyjster}
\title{Test shinyjster}
\usage{
test_shinyjster(
  dir = "apps",
  apps = apps_shinyjster(dir),
  port = 8000,
  host = "127.0.0.1",
  browser = "chrome",
  type = c("serial", "lapply", "parallel", "callr"),
  assert = TRUE
)
}
\arguments{
\item{dir}{base folder to look for applications}

\item{apps}{applications within \verb{dir} to run}

\item{port}{The TCP port that the application should listen on. If the
\code{port} is not specified, and the \code{shiny.port} option is set (with
\code{options(shiny.port = XX)}), then that port will be used. Otherwise,
use a random port.}

\item{host}{The IPv4 address that the application should listen on. Defaults
to the \code{shiny.host} option, if set, or \code{"127.0.0.1"} if not. See
Details.}

\item{browser}{a non-empty character string giving the name of the
    program to be used as the HTML browser.  It should be in the PATH,
    or a full path specified.  Alternatively, an \R function to be
    called to invoke the browser.

    Under Windows \code{NULL} is also allowed (and is the default), and
    implies that the file association mechanism will be used.
  }

\item{type}{Single value to determine how applications are executed. \describe{
\item{\code{'serial'}}{Runs apps one after another using \code{lapply}. \code{port} will be random for each app unless specified.}
\item{\code{'callr'}}{Runs apps using \code{callr::r_bg} using \code{cores} cores. \code{port} will be random for each app to allow concurrent execution.}
\item{\code{'lapply'}}{Runs apps in succession using \code{lapply}. \code{port} will be random for each app unless specified.}
}}

\item{assert}{Logical value which will determine if \code{shinyjster::assert_jster} will be run on the final output. Defaults to \code{TRUE}}
}
\description{
Test shinyjster
}
