% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_shinyjster_app}
\alias{test_shinyjster_app}
\title{Test shinyjster app}
\usage{
test_shinyjster_app(
  name,
  browser_func = switch(name, chrome = shinyjster::selenium_chrome(), firefox =
    shinyjster::selenium_firefox(), edge = shinyjster::selenium_edge(), ie =
    shinyjster::selenium_ie(), {    
    message("unknown browser name supplied. Using 'chrome'")    
    shinyjster::selenium_chrome() }),
  apps = ".."
)
}
\arguments{
\item{name}{browser name to use to determine the browser_func}

\item{browser_func}{One of \code{c("chrome", "firefox", "edge", "ie")}. If an unknown \code{browser_name} is used, \code{browser_func} will default to using \code{chrome}.}

\item{apps}{shiny app to be tested using shinyjster}
}
\description{
Method to avoid editing MANY different files to do the same default behavior of \code{shinyjster::test_jster(browsers = selenium_chrome(), type = "lapply")}.
}
\details{
Some extra logic wraps around whether or not a browser is available on the given platform.
}
\seealso{
\code{\link[=test_shinytest_app]{test_shinytest_app()}} and \code{\link[=test_runtests]{test_runtests()}}
}
