% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_shinytest.R
\name{test_shinytest}
\alias{test_shinytest}
\title{Test shinytest}
\usage{
test_shinytest(
  dir = "apps",
  apps = apps_shinytest(dir, suffix = suffix),
  suffix = platform()
)
}
\arguments{
\item{dir}{Name of directory to look for shiny apps, and to save platform
information.}

\item{apps}{Character vector of shiny applications}

\item{suffix}{An optional suffix for the expected results directory. For
example, if the suffix is \code{"mac"}, the expected directory would be
\code{mytest-expected-mac}.}
}
\description{
Test shinytest
}
