% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_shinytest_app}
\alias{test_shinytest_app}
\title{Test shinytest app}
\usage{
test_shinytest_app(appDir = "..", suffix = shinycoreci::platform())
}
\arguments{
\item{appDir}{Path to the Shiny application to be tested.}

\item{suffix}{An optional suffix for the expected results directory. For
example, if the suffix is \code{"mac"}, the expected directory would be
\code{mytest-expected-mac}.}
}
\description{
Test shinytest app
}
\seealso{
\code{\link[=test_shinyjster_app]{test_shinyjster_app()}} and \code{\link[=test_runtests]{test_runtests()}}
}
