% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app.R
\name{test_shinytest_app}
\alias{test_shinytest_app}
\title{Test shinytest app}
\usage{
test_shinytest_app(appDir = "..", suffix = platform_rversion(), ...)
}
\arguments{
\item{appDir}{Path to directory containing a Shiny app (e.g. \code{app.R}) or
single interactive \code{.Rmd}.}

\item{suffix}{An optional suffix for the expected results directory. For
example, if the suffix is \code{"mac"}, the expected directory would be
\code{mytest-expected-mac}.}

\item{...}{passed to \code{\link[shinytest:testApp]{shinytest::testApp()}}}
}
\description{
Test shinytest app
}
\seealso{
\code{\link[=test_shinyjster_app]{test_shinyjster_app()}}, \code{\link[=test_runtests]{test_runtests()}}, and \code{\link[=platform_rversion]{platform_rversion()}}
}
