% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger.R
\name{trigger}
\alias{trigger}
\alias{trigger_tests}
\alias{trigger_deploy}
\alias{trigger_docker}
\title{Generate a repository event.}
\usage{
trigger(
  event_type,
  repo = "rstudio/shinycoreci",
  client_payload = list(),
  auth_token = Sys.getenv("GITHUB_PAT")
)

trigger_tests(
  repo = "rstudio/shinycoreci",
  auth_token = Sys.getenv("GITHUB_PAT")
)

trigger_deploy(
  repo = "rstudio/shinycoreci",
  auth_token = Sys.getenv("GITHUB_PAT")
)

trigger_docker(
  repo = "rstudio/shinycoreci",
  auth_token = Sys.getenv("GITHUB_PAT")
)
}
\arguments{
\item{event_type}{The name of the event to create on the repository}

\item{repo}{The GitHub repo to create the event on; defaults to
rstudio/shinycoreci}

\item{client_payload}{The JSON object to make available in the workflow as
the \code{github.event.client_payload} object}

\item{auth_token}{Your GitHub \strong{P}ersonal \strong{A}ccess \strong{T}oken; defaults to
\code{Sys.getenv("GITHUB_PAT")}}
}
\description{
This function uses the GitHub API to create a \href{https://developer.github.com/v3/repos/#create-a-repository-dispatch-event}{repository dispatch event}
that can trigger workflows.
}
