% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-test-images.R
\name{view_test_images}
\alias{view_test_images}
\title{View Shinytest Images}
\usage{
view_test_images(repo_dir = ".")
}
\arguments{
\item{repo_dir}{directory to the shinycoreci repo}
}
\description{
View Shinytest Images
}
