% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysinfo.R
\name{write_sysinfo}
\alias{write_sysinfo}
\title{Write system information to a file}
\usage{
write_sysinfo(file = stdout())
}
\arguments{
\item{file}{Name of file, or file obect to write to (defaults to stdout).}
}
\description{
Write system information to a file
}
